Added indicator to tile the player is facing
A red border is now drawn around the tile the player is facing so they can see which tile their tool will interact with before using it.

Moved code to calculate the coordinates of the tile the player is facing to 'get_facing()' in the Player module.
